Add ability to check for container with given name
This adds a new argument, `-container-name`, which will make
check_docker check that a container with the specified name exists and
is running.

This is useful in cases where you use the same container more than once,
but wish to ensure all instances are up. For example, two wordpress
blogs using the container `wordpress` but named `wordpress_A` and
`wordpress_B`.